+++ dgisim/src/character/character.py
@@ -34,6 +34,7 @@ __all__ = [
     "AratakiItto",
     "Bennett",
     "ElectroHypostasis",
+    "Fischl",
     "FatuiPyroAgent",
     "Jean",
     "KaedeharaKazuha",
@@ -804,7 +805,7 @@ class FatuiPyroAgent(Character):
 class Fischl(Character):
     _ELEMENT = Element.ELECTRO
     _WEAPON_TYPE = WeaponType.BOW
-    _TALENT_STATUS = None
+    _TALENT_STATUS = stt.StellarPredatorStatus
     _FACTIONS = frozenset((Faction.MONDSTADT,))
 
     _NORMAL_ATTACK_COST = AbstractDices({
@@ -817,6 +818,69 @@ class Fischl(Character):
     _ELEMENTAL_BURST_COST = AbstractDices({
         Element.ELECTRO: 3,
     })
+
+    @override
+    def _normal_attack(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return normal_attack_template(
+            game_state=game_state,
+            source=source,
+            element=Element.PHYSICAL,
+            damage=2,
+        )
+
+    @override
+    def _elemental_skill1(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.ELECTRO,
+                damage=1,
+                damage_type=DamageType(elemental_skill=True),
+            ),
+            eft.AddSummonEffect(
+                target_pid=source.pid,
+                summon=sm.OzSummon,
+            ),
+        )
+
+    @override
+    def _elemental_burst(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.EnergyDrainEffect(
+                target=source,
+                drain=self.get_max_energy(),
+            ),
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_OFF_FIELD,
+                element=Element.PIERCING,
+                damage=2,
+                damage_type=DamageType(elemental_burst=True),
+            ),
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.ELECTRO,
+                damage=4,
+                damage_type=DamageType(elemental_burst=True),
+            ),
+        )
+
+    @classmethod
+    def from_default(cls, id: int = -1) -> Self:
+        return cls(
+            id=id,
+            alive=True,
+            hp=10,
+            max_hp=10,
+            energy=0,
+            max_energy=3,
+            statuses=stts.Statuses(()),
+            hiddens=stts.Statuses(()),
+            equipments=stts.EquipmentStatuses(()),
+            elemental_aura=ElementalAura.from_default(),
+        )
 
 
 class Jean(Character):
@@ -1239,7 +1303,7 @@ class Keqing(Character):
         )
 
     @classmethod
-    def from_default(cls, id: int = -1) -> Keqing:
+    def from_default(cls, id: int = -1) -> Self:
         return cls(
             id=id,
             alive=True,